+++ src/include/LSP/Workspace.hpp
@@ -67,6 +67,8 @@ public:
 
     /// Whether the file has been marked as ignored by any of the ignored lists in the configuration
     bool isIgnoredFile(const std::filesystem::path& path, const std::optional<ClientConfiguration>& givenConfig = std::nullopt);
+    /// Whether the file has been marked as ignored for auto-importing
+    bool isIgnoredFileForAutoImports(const std::filesystem::path& path, const std::optional<ClientConfiguration>& givenConfig = std::nullopt);
     /// Whether the file has been specified in the configuration as a definitions file
     bool isDefinitionFile(const std::filesystem::path& path, const std::optional<ClientConfiguration>& givenConfig = std::nullopt);