+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - We now apply a datamodel-aware `FindFirstChild` function to the top level datamodel and service types ([#543](https://github.com/JohnnyMorganz/luau-lsp/issues/543))
 - Added autocompletion of children to `:FindFirstChild("")` and `:WaitForChild("")` ([#685](https://github.com/JohnnyMorganz/luau-lsp/issues/685))
 - Attached magic function to `Instance.fromExisting` to allow it to operate similar to `inst:Clone`
+- Added separate configuration `luau-lsp.completion.imports.ignoreGlobs` to filter out files for auto-importing. We no longer check `luau-lsp.ignoreGlobs`.
 
 ### Changed